+++ core/src/main/scala/org/apache/spark/eventhubs/common/rdd/OffsetStoreParams.scala
@@ -15,21 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.spark.eventhubscommon
+package org.apache.spark.eventhubs.common.rdd
 
-/**
- * interface representing the bridge between EventHubs and Spark-side processing engine
- * (Direct DStream or Structured Streaming)
- */
-private[spark] trait EventHubsConnector {
-
-  // the id of the stream which is mapped from eventhubs instance
-  def streamId: Int
-
-  // uniquely identify the entities in eventhubs side, it can be the EventHubs namespace or the
-  // combination of namespace and eventhubs name and streamId
-  def uid: String
-
-  // the list of eventhubs partitions connecting with this connector
-  def connectedInstances: List[EventHubNameAndPartition]
-}
+// a helper object to avoid serialzing offsetstore instances
+private[spark] case class OffsetStoreParams(checkpointDir: String,
+                                            streamId: Int,
+                                            uid: String,
+                                            subDirs: String*)
