+++ core/src/test/scala/org/apache/spark/eventhubs/common/client/EventHubsClientWrapperSuite.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.eventhubscommon.client
+package org.apache.spark.eventhubs.common.client
 
 import org.scalatest.{ BeforeAndAfter, FunSuite }
 import org.scalatest.mock.MockitoSugar
