+++ core/src/test/scala/org/apache/spark/sql/streaming/eventhubs/EventHubsStreamTest.scala
@@ -25,7 +25,6 @@ import scala.language.experimental.macros
 import scala.reflect.ClassTag
 import scala.util.Random
 import scala.util.control.NonFatal
-
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{ FileSystem, Path }
 import org.scalatest.{ Assertions, BeforeAndAfter }
@@ -35,13 +34,12 @@ import org.scalatest.concurrent.PatienceConfiguration.Timeout
 import org.scalatest.exceptions.TestFailedDueToTimeoutException
 import org.scalatest.time.Span
 import org.scalatest.time.SpanSugar._
-
 import org.apache.spark.DebugFilesystem
-import org.apache.spark.eventhubscommon.EventHubsConnector
-import org.apache.spark.eventhubscommon.progress.ProgressTrackerBase
-import org.apache.spark.eventhubscommon.utils._
+import org.apache.spark.eventhubs.common.EventHubsConnector
+import org.apache.spark.eventhubs.common.progress.ProgressTrackerBase
+import org.apache.spark.eventhubs.common.utils._
 import org.apache.spark.sql.{ Dataset, Encoder, QueryTest, Row }
-import org.apache.spark.sql.catalyst.encoders.{ encoderFor, ExpressionEncoder, RowEncoder }
+import org.apache.spark.sql.catalyst.encoders.{ ExpressionEncoder, RowEncoder, encoderFor }
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.catalyst.util._
 import org.apache.spark.sql.execution.streaming._
