+++ core/src/test/scala/org/apache/spark/streaming/eventhubs/CheckpointAndProgressTrackerTestSuiteBase.scala
@@ -18,7 +18,7 @@
 package org.apache.spark.streaming.eventhubs
 
 import org.apache.hadoop.fs.{ Path, PathFilter }
-import org.apache.spark.eventhubscommon.OffsetRecord
+import org.apache.spark.eventhubs.common.OffsetRecord
 import org.apache.spark.streaming._
 import org.apache.spark.streaming.dstream.DStream
 import org.apache.spark.streaming.eventhubs.checkpoint.DirectDStreamProgressTracker
