+++ core/src/test/scala/org/apache/spark/streaming/eventhubs/SharedUtils.scala
@@ -21,11 +21,11 @@ import java.nio.file.Files
 
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{ FileSystem, Path }
+import org.apache.spark.eventhubs.common.EventHubsConnector
 import org.scalatest.{ BeforeAndAfterEach, FunSuite }
 import org.apache.spark.{ SparkConf, SparkContext }
-import org.apache.spark.eventhubscommon.EventHubsConnector
-import org.apache.spark.eventhubscommon.client.EventHubsClientWrapper
-import org.apache.spark.eventhubscommon.progress.ProgressTrackerBase
+import org.apache.spark.eventhubs.common.client.EventHubsClientWrapper
+import org.apache.spark.eventhubs.common.progress.ProgressTrackerBase
 import org.apache.spark.streaming.{ Duration, Seconds, StreamingContext }
 import org.apache.spark.streaming.eventhubs.checkpoint.{
   DirectDStreamProgressTracker,
